+++ Packages/GoSublime/src/gosubli.me/something-borrowed/gocode/gosublime_gocode_bridge.go
@@ -1,6 +1,9 @@
 package gocode
 
 import (
+	"fmt"
+	"go/build"
+	"os"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -16,13 +19,36 @@ type MargoConfig struct {
 	GOPATHS       []string
 }
 
+type margoEnv struct {
+	LibPath       string
+	GOOS          string
+	GOARCH        string
+	Compiler      string
+	GOROOT        string
+	GOPATH        string
+	InstallSuffix string
+}
+
+func (m *margoEnv) assign(c *config, p *package_lookup_context) {
+	c.LibPath = m.LibPath
+
+	p.GOOS = m.GOOS
+	p.GOARCH = m.GOARCH
+	p.Compiler = m.Compiler
+	p.GOROOT = m.GOROOT
+	p.GOPATH = m.GOPATH
+	p.InstallSuffix = m.InstallSuffix
+}
+
 type margoState struct {
 	sync.Mutex
 
 	ctx       *auto_complete_context
-	env       *gocode_env
+	env       *package_lookup_context
 	pkgCache  package_cache
 	declCache *decl_cache
+	prevEnv   margoEnv
+	prevPkg   string
 }
 
 type MargoCandidate struct {
@@ -32,7 +58,7 @@ type MargoCandidate struct {
 }
 
 func newMargoState() *margoState {
-	env := &gocode_env{}
+	env := &package_lookup_context{}
 	pkgCache := new_package_cache()
 	declCache := new_decl_cache(env)
 	return &margoState{
@@ -47,7 +73,7 @@ func (m *margoState) Complete(c MargoConfig, file []byte, filename string, curso
 	m.Lock()
 	defer m.Unlock()
 
-	m.updateConfig(c)
+	m.updateConfig(c, filename)
 
 	list, _ := m.ctx.apropos(file, filename, cursor)
 	candidates := make([]MargoCandidate, len(list))
@@ -61,7 +87,7 @@ func (m *margoState) Complete(c MargoConfig, file []byte, filename string, curso
 	return candidates
 }
 
-func (m *margoState) updateConfig(c MargoConfig) {
+func (m *margoState) updateConfig(c MargoConfig, filename string) {
 	pl := []string{}
 	osArch := runtime.GOOS + "_" + runtime.GOARCH
 	if c.InstallSuffix != "" {
@@ -78,6 +104,30 @@ func (m *margoState) updateConfig(c MargoConfig) {
 		add(p)
 	}
 
+	sep := string(filepath.ListSeparator)
+
+	nv := margoEnv{
+		LibPath:       strings.Join(pl, sep),
+		GOOS:          runtime.GOOS,
+		GOARCH:        runtime.GOARCH,
+		Compiler:      runtime.Compiler,
+		GOROOT:        c.GOROOT,
+		GOPATH:        strings.Join(c.GOPATHS, sep),
+		InstallSuffix: c.InstallSuffix,
+	}
+	nv.assign(&g_config, m.env)
 	g_config.ProposeBuiltins = c.Builtins
-	g_config.LibPath = strings.Join(pl, string(filepath.ListSeparator))
+	m.env.CurrentPackagePath = ""
+	p, _ := m.env.ImportDir(filepath.Dir(filename), build.FindOnly)
+	if p != nil {
+		m.env.CurrentPackagePath = p.ImportPath
+	}
+	if m.prevPkg != m.env.CurrentPackagePath {
+		m.prevPkg = m.env.CurrentPackagePath
+		fmt.Fprintf(os.Stderr, "Gocode pkg: %#v\n", m.env.CurrentPackagePath)
+	}
+	if m.prevEnv != nv {
+		m.prevEnv = nv
+		fmt.Fprintf(os.Stderr, "Gocode env: %#v\n", nv)
+	}
 }